using new interaction code as main entry point
