+++ impsy/impsy.py
@@ -5,7 +5,6 @@ import click
 from .dataset import dataset
 from .train import train
 from .interaction import run
-from .interaction_config import start_genai_module
 from .tests import test_mdrnn
 
 @click.group()
@@ -18,7 +17,6 @@ def main():
     cli.add_command(dataset)
     cli.add_command(train)
     cli.add_command(run)
-    cli.add_command(start_genai_module)
     cli.add_command(test_mdrnn)
     # runs the command line interface
     cli()
